Exercise 1 - World, Glue, Pointer
This first commit include the basic world, a script that handle the
controller in order to display it in front of the camera and a pointer
script taking into consideration the orientation of the Daydream
controller to apply it to the VR Controller. The VR COntroller cast a
ray and display a pointerDot object were the ray is casted.